Flask/requirements.txt
Flask>=3.0
requests>=2.32
python-dotenv>=1.0
psycopg[binary]>=3.2